chore(analytics): set analytics codes for the generated docs
This is done as a separate step from the previous commit to help with
rebases.  In my most recent rebase, the conflict was:

  Our version:
      1. client/asset/config.yaml.sample was renamed to client/asset/config.yaml
      2. AND modified to include the new keys
  Dart team's version:
      1. client/asset/config.yaml.sample was moved to
      client/lib/config/config.yaml.sample

It is easy to make a mistake during this merge conflict.  So I've split
out the rename and the update into two commits.  Naomi's commit now only
does the rename.  This commit updates the renamed file.